Make the finish template create a logfile
Added a couple of exec statements to finish.erb to make all of the output the script generates go to /root/finish_script.log. This makes it a *lot* easier to figure out why a deployment fails.

It would be even nicer to get the logging output into The Foreman itself, but this gives you at least a starting point for troubleshooting.